Docker based build flow
Revamped the build process to be Docker based and leveraging Concourse for
automation.

Still TBD is documenting and some helpers around the developer workflow. It is
automated and functional, but not locally usable for those developing. That
will be updated shortly.

This also includes a number of small updates, including moving to Buildroot
2015.08.1, ACI generation with acbuild, updates to grub-install.sh for a
container environment, and more.