Update to 4.2.5 kernel. Simplify config file. Added stripping modules.
Update to the 4.2.5 kernel.

Move from checking in the full kernel config file to just the defconfig.
This can be generated with "make savedefconfig" and reloaded with
"make olddefconfig".

Also added the parameter to have "--strip-unneeded" used on the modules.
Try and save some space, and seems the safest option.