Add in overlayfs patch to minimize overlayfs xattr buffer size
This is a change authored by Vito Caputo of CoreOS. It minimizes the size of
the buffer allocated in ovl_copy_xattr when copying up a file. It had been
observed on some systems that modifying files in a overlayfs would fail
with memory allocation errors, even when memory was available but the memory
buffer was full.