This updates disk image settings to handle A/B deploys/updates.
This changes the disk partition structure to have A/B partitions which use GPT
priority settings. The grub.cfg is updated to detect the priority and decide
wether to boot an A or a B kernel/initrd image.

The partitions are essentially empty, but just used for storing metadata.

The size of the boot partition has been increased to 256mb to have more leway
for the size of the images.

A new ACI build has been added for kurma-upgrader, which is used to update the
kernel and kurma daemon on a host and trigger a reboot.