Updated the ACIs with the linux namespaces isolator to be expanded.
The expanded isolator support will eventually be expanded to support a
container sharing a namespace with another. This is the schema related
changes to enable that.

Also, the kurma-init processes default namespaces have been changed to
just pid and mount. This removes IPC and UTS, as some base images will
share these namespaces with the host now.