Add instrumentation for networking plugins.
Adds the building of container images for the default networking
plugins. This includes our own plugin for setting up the loopback
device, as well as a container image for the Container Network
Interface (CNI) networking plugins.

The default Kurma configuration will now instrument a loopback and a
veth device on a bridge, with ip forwarding configured to the host's
interface. This is roughly inline with default behavior for Docker.